+++ src/main/java/com/example/firstproject/services/IBankService.java
@@ -1,0 +1,15 @@
+package com.example.firstproject.services;
+
+import com.example.firstproject.models.CompteDto;
+import com.example.firstproject.models.OperationCompteDto;
+
+public interface IBankService {
+
+    CompteDto creerCompte(CompteDto compteDto);
+
+    CompteDto tranferer(String numeroCompteExpediteur, String numeroCompteDestinataire, Double montantTransfert);
+
+    CompteDto crediter(OperationCompteDto operationCompteDto);
+    CompteDto debiter (OperationCompteDto operationCompteDto);
+   // CompteDto crediterOuDebiter(OperationCompteDto operationCompteDto);
+}
